requirements.txt
aiohttp==3.6.2
aiofiles==0.4.0
Babel==2.8.0
beautifulsoup4==4.8.2
dateparser==0.7.2
Flask
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-datastore==1.10.0
google-cloud-logging==1.14.0
grpcio==1.26.0
opencensus==0.7.6
opencensus-ext-flask==0.7.3
opencensus-ext-zipkin==0.2.2
opencensus-ext-stackdriver==0.7.2
prometheus_client==0.7.1
python-dateutil
pytz==2019.3

asynctest
coverage
mypy
black
faker
isort
PyHamcrest
pylint
pytest
pytest-asyncio
pytest-cov
watchdog

